refactor(lang): extract prettify per-rule formatters into separate files
Per-rule formatters become free functions taking (f: Formatter, ctx)
instead of methods on the Formatter class. Each rule lives in its own
file, the leaf walker lives in its own file, and index.ts shrinks from
~1180 lines to ~145.

  formatter.ts        Formatter class (state + format() dispatcher)
  leaf.ts             emitVisibleToken + per-token state mutators
                      (note, flushHiddenBefore, startStatementLine, ...)
                      and read-only helpers (approxInlineSpan,
                      hasCommentsInRange, formatTokenRange).
  inline-renderer.ts  renderItemInline (flat-string mirror).
  block-body.ts       formatBlockBody, formatTopLevel.
  sections.ts         formatSectionStatement / formatSectionList.
  field-properties.ts formatFieldProperties (postfix `{...}`).
  pick-case.ts        formatPickStatement, formatCaseStatement.
  binary-chain.ts     formatBinaryChain.
  index.ts            prettify() entry + type re-exports.

Per-rule files type-import Formatter from formatter.ts -- no runtime
cycles. The dispatcher (the only method left on Formatter) calls into
the per-rule files; per-rule files call f.format(child) for recursion.

Also drops two dead methods (emitVerbatim, formatToString) the original
class kept around as "tools" but nothing called.

No behavior change. 55/55 prettifier tests still pass.
